requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
sqlalchemy==2.0.45
alembic==1.18.1
jinja2==3.1.6
python-dotenv==1.2.1
aiosqlite==0.22.1
pydantic-settings==2.12.0
httpx==0.28.1
pytest==9.0.2
pytest-asyncio==1.3.0
pyjwt==2.10.1
email-validator==2.3.0
resend==2.21.0
gunicorn
ruff
pre-commit
